rework StateManager
