+++ jabgui/src/test/java/org/jabref/gui/externalfiles/FindUnlinkedFilesActionTest.java
@@ -4,6 +4,7 @@ import java.nio.file.Path;
 import java.util.Optional;
 
 import org.jabref.gui.DialogService;
+import org.jabref.gui.JabRefGuiStateManager;
 import org.jabref.gui.StateManager;
 import org.jabref.logic.shared.DatabaseLocation;
 import org.jabref.model.database.BibDatabaseContext;
@@ -31,7 +32,7 @@ class FindUnlinkedFilesActionTest {
     @BeforeEach
     void setUp() {
         MockitoAnnotations.openMocks(this);
-        stateManager = new StateManager();
+        stateManager = new JabRefGuiStateManager();
         action = new FindUnlinkedFilesAction(dialogService, stateManager);
     }
