+++ jabgui/src/test/java/org/jabref/gui/search/GetLastSearchHistoryTest.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javafx.stage.Stage;
 
+import org.jabref.gui.JabRefGuiStateManager;
 import org.jabref.gui.StateManager;
 import org.jabref.model.database.BibDatabaseContext;
 
@@ -16,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(ApplicationExtension.class)
 class GetLastSearchHistoryTest {
-    private final StateManager stateManager = new StateManager();
+    private final StateManager stateManager = new JabRefGuiStateManager();
     private final BibDatabaseContext dbContext1 = new BibDatabaseContext();
     private final BibDatabaseContext dbContext2 = new BibDatabaseContext();
