+++ jabgui/src/test/java/org/jabref/gui/util/OpenConsoleActionTest.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.frame.OpenConsoleAction;
 import org.jabref.gui.preferences.GuiPreferences;
+import org.jabref.logic.util.OptionalObjectProperty;
 import org.jabref.model.database.BibDatabaseContext;
 
 import org.junit.jupiter.api.BeforeEach;
