+++ jablib/src/main/java/org/jabref/logic/util/OptionalObjectProperty.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.util;
+package org.jabref.logic.util;
 
 import java.util.Optional;
