+++ jabsrv/src/main/java/org/jabref/http/CliStateManager.java
@@ -1,0 +1,28 @@
+package org.jabref.http;
+
+import java.util.List;
+import java.util.Optional;
+
+import javafx.collections.ObservableList;
+
+import org.jabref.logic.search.IndexManager;
+import org.jabref.logic.util.OptionalObjectProperty;
+import org.jabref.model.database.BibDatabaseContext;
+import org.jabref.model.entry.BibEntry;
+
+public interface CliStateManager {
+
+    ObservableList<BibDatabaseContext> getOpenDatabases();
+
+    OptionalObjectProperty<BibDatabaseContext> activeDatabaseProperty();
+
+    ObservableList<BibEntry> getSelectedEntries();
+
+    Optional<IndexManager> getIndexManager(BibDatabaseContext database);
+
+    Optional<BibDatabaseContext> getActiveDatabase();
+
+    List<String> collectAllDatabasePaths();
+
+    boolean isRunningInCli();
+}
