+++ jabsrv/src/main/java/org/jabref/http/server/LibrariesResource.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
+import org.jabref.http.CliStateManager;
 import org.jabref.http.server.services.FilesToServe;
-import org.jabref.http.server.services.GuiBridge;
 import org.jabref.logic.util.io.BackupFileUtil;
 
 import com.google.gson.Gson;
@@ -19,7 +19,7 @@ import jakarta.ws.rs.core.MediaType;
 public class LibrariesResource {
 
     @Inject
-    private GuiBridge guiBridge;
+    private CliStateManager cliStateManager;
 
     @Inject
     private FilesToServe filesToServe;
@@ -34,9 +34,9 @@ public class LibrariesResource {
         if (!filesToServe.isEmpty()) {
             pathStream = filesToServe.getFilesToServe().stream();
         } else {
-            pathStream = guiBridge.getOpenDatabases().stream()
-                                                       .filter(context -> context.getDatabasePath().isPresent())
-                                                       .map(context -> context.getDatabasePath().get());
+            pathStream = cliStateManager.getOpenDatabases().stream()
+                                        .filter(context -> context.getDatabasePath().isPresent())
+                                        .map(context -> context.getDatabasePath().get());
         }
         List<String> fileNamesWithUniqueSuffix = pathStream.map(path -> path.getFileName() + "-" + BackupFileUtil.getUniqueFilePrefix(path))
                                                            .toList();
