+++ jabsrv/src/main/java/org/jabref/http/server/LibraryResource.java
@@ -7,11 +7,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+import org.jabref.http.CliStateManager;
 import org.jabref.http.JabrefMediaType;
 import org.jabref.http.dto.BibEntryDTO;
 import org.jabref.http.dto.LinkedPdfFileDTO;
 import org.jabref.http.server.services.FilesToServe;
-import org.jabref.http.server.services.GuiBridge;
 import org.jabref.http.server.services.ServerUtils;
 import org.jabref.logic.citationstyle.JabRefItemDataProvider;
 import org.jabref.logic.preferences.CliPreferences;
@@ -48,7 +48,7 @@ public class LibraryResource {
     CliPreferences preferences;
 
     @Inject
-    GuiBridge guiBridge;
+    CliStateManager cliStateManager;
 
     @Inject
     FilesToServe filesToServe;
@@ -169,7 +169,7 @@ public class LibraryResource {
                            .build();
         }
 
-        java.nio.file.Path library = ServerUtils.getLibraryPath(id, filesToServe, guiBridge);
+        java.nio.file.Path library = ServerUtils.getLibraryPath(id, filesToServe, cliStateManager);
         String libraryAsString;
         try {
             libraryAsString = Files.readString(library);
@@ -184,7 +184,7 @@ public class LibraryResource {
     }
 
     private java.nio.file.Path getJabMapPath(String id) {
-        java.nio.file.Path libraryPath = ServerUtils.getLibraryPath(id, filesToServe, guiBridge);
+        java.nio.file.Path libraryPath = ServerUtils.getLibraryPath(id, filesToServe, cliStateManager);
         String newName = libraryPath.getFileName().toString().replaceFirst("\\.bib$", ".jmp");
         return libraryPath.getParent().resolve(newName);
     }
@@ -199,7 +199,7 @@ public class LibraryResource {
      * @param id - also "demo" for the Chocolate.bib file
      */
     private BibDatabaseContext getDatabaseContext(String id) throws IOException {
-        return ServerUtils.getBibDatabaseContext(id, filesToServe, guiBridge, preferences.getImportFormatPreferences());
+        return ServerUtils.getBibDatabaseContext(id, filesToServe, cliStateManager, preferences.getImportFormatPreferences());
     }
 
     /**
